requirements.txt
psutil>=5.9.0
pyyaml>=6.0.1
numpy>=1.24.0
asyncio>=3.4.3
typing>=3.7.4.3
dataclasses>=0.6
PyYAML>=6.0.1
cmake>=3.25.0 